Work on client test coverage
- PingPong now closes after reporting the disconnection, no longer sets the TimeoutStrategy null and checks whether it is already closed (if close is called again)

- Added a test for the interaction between the server and the client

- Client now parses the dish menu using ExternalDishMenu parts (instead of FileDishMenu parts) (was causing a bug)

- DishMenuHelper now has methods to parse serialised dish menus coming from the External

- Fixed the DummyServiceConnectionManager and DummyExternalConnector needing manual setting of the connection target. DummyServiceConnectionManager now has another boolean, which (when the attribute is true) attempts to reconnect upon disconnection, if the connection is allowed.

+++ Server/sel.sales/src/main/java/model/filewriter/FileAccess.java
@@ -57,7 +57,10 @@ public abstract class FileAccess implements IFileAccess {
 	public boolean deleteFile() {
 		if (this.fileExists()) {
 			this.closeRAF();
-			boolean isDeleted = this.activeFile.delete();
+			boolean isDeleted = false;
+			if (this.activeFile != null) {
+				isDeleted = this.activeFile.delete();
+			}
 			this.activeFile = null;
 			return isDeleted;
 		}
